+++ packages/server/jest.config.ts
@@ -1,6 +1,7 @@
 import type { Config } from 'jest'
 
 const config: Config = {
+  preset: 'ts-jest',
   setupFiles: ['./test/setup.ts'],
 
   moduleNameMapper: {