+++ packages/server/src/notifier/model/email/index.ts
@@ -1,13 +1,18 @@
 import { pick } from 'lodash'
-import { verbose } from 'npmlog'
+import { verbose, error } from 'npmlog'
 
-import { Email } from '@/common/utils/email'
+import { Email, EmailProvider } from '@/common/utils/email'
 
+import {
+  fromElectionAnnouncingStartedNotification,
+  fromElectionRevealingStartedNotification,
+  fromElectionVotingStartedNotification,
+} from './election'
 import { fromPostAddedNotification, fromThreadCreatedNotification } from './forum'
 import { Notification, hasEmailAddress } from './utils'
 
-export const notifyByEmail =
-  (send: (email: Email) => Promise<void>) =>
+export const createEmailNotifier =
+  (emailProvider: EmailProvider) =>
   async (notification: Notification): Promise<void> => {
     if (!hasEmailAddress(notification)) throw Error(`No email address defined for member ${notification.memberId}`)
 
@@ -17,11 +22,28 @@ export const notifyByEmail =
       pick(notification, 'id', 'eventId', 'kind', 'entityId')
     )
 
-    // The functions should be chained here eg.
-    // fromPostAddedNotification(notification) ?? fromProposalPostCreated(notification) ?? ...
-    const email = await (fromPostAddedNotification(notification) ?? fromThreadCreatedNotification(notification))
+    const emailHandlers = [
+      fromPostAddedNotification,
+      fromThreadCreatedNotification,
+      fromElectionAnnouncingStartedNotification,
+      fromElectionVotingStartedNotification,
+      fromElectionRevealingStartedNotification,
+    ]
+    const emailPromises = emailHandlers.map((handler) => handler(notification))
+    const emailResults = await Promise.all(emailPromises)
+
+    let email: Email | undefined
+    for (const result of emailResults) {
+      if (result) {
+        if (email) {
+          error('Email notification', 'Multiple emails generated for notification', notification)
+        } else {
+          email = result
+        }
+      }
+    }
 
     if (!email) throw Error(`No email template found for notification ${notification.kind}`)
 
-    await send(email)
+    await emailProvider.sendEmail(email)
   }