+++ packages/ui/src/common/components/MarkdownPreview/MarkdownPreviewStyles.tsx
@@ -86,7 +86,7 @@ export const MarkdownPreviewStyles = createGlobalStyle<MarkdownPreviewStylesProp
     font-size: ${normalFontSize};
     line-height: 24px;
     font-weight: 400;
-    color: ${Colors.Black[500]};
+    color: ${normalColor};
   }
   
   .markdown-preview li + li {
@@ -108,7 +108,7 @@ export const MarkdownPreviewStyles = createGlobalStyle<MarkdownPreviewStylesProp
     counter-reset: ol-list-counter;
   }
 
-  .markdown-preview ol li {
+  .markdown-preview ol > li {
     counter-increment: ol-list-counter;
   }
   
@@ -116,7 +116,7 @@ export const MarkdownPreviewStyles = createGlobalStyle<MarkdownPreviewStylesProp
     margin-top: 0;
   }
 
-  .markdown-preview ol li::before {
+  .markdown-preview ol > li::before {
     content: counter(ol-list-counter)'.';
     position: absolute;
     left: -20px;