+++ packages/ui/src/mocks/data/members.ts
@@ -1,11 +1,12 @@
-import { MemberFieldsFragment } from '@/memberships/queries'
+import { MemberWithDetailsFieldsFragment } from '@/memberships/queries'
 
 import rawMembers from './raw/members.json'
 
-export type Membership = Omit<MemberFieldsFragment, '__typename'>
+export type Membership = Omit<MemberWithDetailsFieldsFragment, '__typename'>
 export const member = (handle: string, { roles = [], ...extra }: Partial<Membership> = {}) =>
   ({
     ...rawMembers.find((member) => member.handle === handle),
+    invitees: [],
     ...extra,
     roles,
   } as Membership)