+++ packages/ui/src/mocks/helpers/asChainData.ts
@@ -1,22 +1,23 @@
+import { createType } from '@joystream/types'
 import { mapValues } from 'lodash'
-
-import { asBN } from '@/common/utils'
 
 export const asChainData = (data: any): any => {
   switch (Object.getPrototypeOf(data).constructor.name) {
     case 'Object':
-      return mapValues(data, asChainData)
+      return withUnwrap(mapValues(data, asChainData))
 
     case 'Array':
       return data.map(asChainData)
 
     case 'Number':
-      return asBN(data)
+      return createType('u128', data)
 
     case 'String':
-      return isNaN(data) ? data : asBN(data)
+      return isNaN(data) ? data : createType('u128', data)
 
     default:
       return data
   }
 }
+
+const withUnwrap = (data: Record<any, any>) => Object.defineProperty(data, 'unwrap', { value: () => data })