+++ packages/ui/src/mocks/providers/index.tsx
@@ -4,12 +4,18 @@ import React, { useMemo } from 'react'
 
 import { MockAccountsProps, MockAccountsProvider } from './accounts'
 import { MockApiProvider, MockApiProps } from './api'
-import { MockQNProps, MockQNProvider } from './query-node'
+import { MockBackendProps, MockBackendProvider } from './backend'
+import { MockGqlProps, MockGqlProvider } from './gql'
+import { MockLocalStorage, useMockLocalStorage } from './useMockLocalStorage'
 
-export type MocksParameters = MockApiProps & MockQNProps & MockAccountsProps
+export * from './router'
+
+export type MocksParameters = MockApiProps & { gql?: MockGqlProps } & MockBackendProps &
+  MockAccountsProps &
+  MockLocalStorage
 
 type Context = StoryContext & {
-  parameters: { mocks: MocksParameters | ((storyContext: StoryContext) => MocksParameters) }
+  parameters: { mocks?: MocksParameters | ((storyContext: StoryContext) => MocksParameters) }
 }
 
 export const MockProvidersDecorator = (Story: CallableFunction, storyContext: Context) => {
@@ -18,13 +24,17 @@ export const MockProvidersDecorator = (Story: CallableFunction, storyContext: Co
     return isFunction(mocks) ? mocks(storyContext) : mocks
   }, [storyContext])
 
+  useMockLocalStorage(mocks?.localStorage)
+
   return (
     <MockApiProvider chain={mocks?.chain}>
-      <MockQNProvider queryNode={mocks?.queryNode}>
+      <MockGqlProvider queries={mocks?.gql?.queries} mutations={mocks?.gql?.mutations}>
         <MockAccountsProvider accounts={mocks?.accounts}>
-          <Story />
+          <MockBackendProvider backend={mocks?.backend}>
+            <Story />
+          </MockBackendProvider>
         </MockAccountsProvider>
-      </MockQNProvider>
+      </MockGqlProvider>
     </MockApiProvider>
   )
 }